refactor: ThumbnailJobPayload dto 패키지로 이동
