+++ nmnb-webflux/src/main/kotlin/nmnb/webflux/global/infrastructure/external/redis/dto/ThumbnailJobPayload.kt
@@ -1,4 +1,4 @@
-package nmnb.webflux.global.infrastructure.external.redis
+package nmnb.webflux.global.infrastructure.external.redis.dto
 
 data class ThumbnailJobPayload(
     val postId: Long,
